requirements.txt
# Django 和相關依賴
Django==4.2
django-ninja
django-cors-headers
django-redis
djangorestframework-simplejwt

# 工具
python-dotenv
requests

# 監控與除錯
sentry-sdk
django-silk

# 圖片處理
Pillow

# 資料庫
psycopg2-binary